build: enforce the warning promise on Linux, harden registration (D30, E4, E5)
E5 — the MSVC branch has /WX and the GCC/Clang branch had nothing, so Linux
and macOS emitted 1,636 warnings and ignored every one while AGENTS.md
promised "compiles without warnings on Windows and Linux". Followed the row's
own recommendation: narrow the set so the promise matches what is enforced,
then add -Werror.

Dropped exactly the three the measurement found in bulk — 1,447 of the 1,636:
-Wold-style-cast (almost all (SQLCHAR*)"literal" at ODBC call sites),
-Wsign-conversion (SQLSMALLINT/SQLULEN/size_t traffic inherent to the API),
-Wuseless-cast. E5 triaged the remainder and found no live defect, so this is
not switching off what fails: a warning nobody can act on is noise, and the
noise is what let the other 189 hide.

Fixing the 13 sites in the way turned up four pieces of genuinely dead code.
Three files carried an anonymous-namespace first_sqlstate that nothing calls
— the call sites resolve to TestBase's static member, and boundary_tests.cpp
even says "a stand-in that C5 deletes", which C5 then missed. Plus hex_bytes,
orphaned four commits ago by my own D64, and escape_sequence_tests.cpp's
match lambda from C6. Reference run 193/193 unchanged: dead, not merely
unused.

D30 — all four registration defects share a shape: the script prints success
while the registration will not be used. register.sh reinstalls rather than
letting odbcinst bump UsageCount and leave a stale Driver=. register.ps1 no
longer "succeeds" into HKCU, which the Windows DM never reads for drivers, or
into WOW6432Node from a 32-bit shell. Its dev-layout probe finds Debug as
well as Release. Both scripts now read the registration back and fail if it
is not the binary asked for. Verified by running both paths.

E4 — already fixed, and the row's line reference is stale: CONTRIBUTING.md is
103 lines now and already documents the Ninja configure. Confirmed rather
than assumed by running it — the VS tree has no compile_commands.json, the
Ninja one has 61 entries. Found and closed one real gap: build-tidy/ was not
gitignored, so following the docs left an untracked tree.

+++ src/reporting/json_reporter.cpp
@@ -106,7 +106,8 @@ void JsonReporter::report_summary(size_t total_tests, size_t passed, size_t fail
     summary["total_duration_us"] = total_duration.count();
 
     if (scored > 0) {
-        summary["pass_rate"] = (passed * 100.0) / scored;
+        summary["pass_rate"] =
+            (static_cast<double>(passed) * 100.0) / static_cast<double>(scored);
     } else {
         summary["pass_rate"] = 0.0;
     }
